Buffs added

+++ Assets/_Game/Scripts/Data/CharacterSO.cs
@@ -12,7 +12,7 @@ namespace OceanGame
         [field: SerializeField] public int BaseDefense { get; private set; }
         [field: SerializeField] public int BaseDamage { get; private set; }
         [field: SerializeField, Range(0, 100)] public int KnockbackForce { get; private set; }
-        [field: SerializeField] public float BaseSpeed { get; private set; }
+        [field: SerializeField] public int BaseSpeed { get; private set; }
         [field: SerializeField] public float BaseIFrameDuration { get; private set; }
         [field: SerializeField] public float BaseTurnSharpness { get; private set; }
         [field: SerializeField, Range(0f, 1f)] public float BaseKbResist { get; private set; }
